added a novel test for a new source UrlSource to demonstrate plugin functionality of Sources/Sinks, and wrote some documenation
