+++ test/sources/test_url.py
@@ -1,28 +1,6 @@
-from argschema.sources import ArgSource
-from argschema.schemas import DefaultSchema
-from argschema.fields import Str,Int
-from argschema import ArgSchemaParser
 import requests 
 import mock
-from test_classes import MySchema
-
-class UrlSourceConfig(DefaultSchema):
-    input_host = Str(required=True, description="host of url")
-    input_port = Int(required=False, default=80, description="port of url")
-    input_url = Str(required=True, description="location on host of input")
-
-class UrlSource(ArgSource):
-    ConfigSchema = UrlSourceConfig
-    def get_dict(self):
-        url = "http://{}:{}/{}".format(self.input_host,
-                                self.input_port,
-                                self.input_url)
-        response = requests.get(url)
-        return response.json()
-
-class UrlArgSchemaParser(ArgSchemaParser):
-    default_configurable_sources = [UrlSource]
-    default_schema = MySchema
+from url_source import UrlArgSchemaParser, UrlSource
 
 # This method will be used by the mock to replace requests.get
 def mocked_requests_get(*args, **kwargs):
